pyproject.toml
[tool.poetry]
name = "discursus-data-platform"
version = "0.1.0"
description = "The discursus project is an OSS social analytics data product that continuously delivers data on protest movements throughout North America."
authors = ["Olivier Dupuis <odupuis@lantrns.co>"]
license = "MIT"
packages = [{include = "discursus_data_platform"}]

[tool.poetry.dependencies]
python = ">=3.8,<3.11"
sqlalchemy = "1.4.46"
future = "^0.18.2"
dbt-core = "1.4.5"
dbt-snowflake = "1.4.2"
dagster = "1.3.5"
dagit = "1.3.5"
dagster-graphql = "^1.3.5"
dagster-postgres = "0.19.5"
dagster-docker = "0.19.5"
dagster-shell = "0.19.5"
dagster-pandas = "0.19.5"
dagster-dbt = "0.19.5"
dagster-aws = "0.19.5"
dagster-snowflake = "0.19.5"
dagster-hex = "0.1.2"
pandas = "^1.5.2"
google-cloud-bigquery = "^3.10.0"
db-dtypes = "^1.1.1"
optparse-pretty = "^0.1.1"
boto3 = "^1.26.35"
spacy = "^3.5.2"
en_core_web_sm = { url = "https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0.tar.gz" }
fsspec = "^2022.11.0"
saf-aws = {git = "https://github.com/discursus-data/saf_aws", rev = "release/0.2"}
saf-gdelt = {git = "https://github.com/discursus-data/saf_gdelt", rev = "release/0.2.1"}
saf-web-scraper = {git = "https://github.com/discursus-data/saf_web_scraper", rev = "release/0.3.1"}
saf-openai = {git = "https://github.com/discursus-data/saf_openai", rev = "release/0.1"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"